🧪add test save-custom files test.ts

+++ test/domain/use-cases/save-file.use-case.test.ts
@@ -5,8 +5,10 @@ import { SaveFile } from '../../../src/domain/use-cases/save-file.use-case';
 
 describe('SaveFileUseCase', () => {
 
+    
     afterEach(()=>{
-        fs.rmSync('outputs', { recursive: true });
+        if (fs.existsSync('outputs')) fs.rmSync('outputs', { recursive: true });
+        if (fs.existsSync('custom-outputs')) fs.rmSync('custom-outputs', { recursive: true });
     });
 
 
@@ -29,6 +31,25 @@ describe('SaveFileUseCase', () => {
 
     test('should save file with custom values', () => {
 
+        const saveFile = new SaveFile();
+        const options = {
+            fileContent: 'custom content',
+            fileDestination: 'custom-outputs/file-destination',
+            fileName:  'custom-table-name'
+        }
+
+        const { fileContent, fileDestination, fileName } = options;
+
+        const filePath = `${fileDestination}/${fileName}.txt`;
+
+        
+        const result = saveFile.execute(options);
+        const fileExists = fs.existsSync(filePath); 
+        const fileContentRead = fs.readFileSync(filePath, { encoding: 'utf-8' })
+
+        expect( result ).toBe( true );
+        expect( fileExists ).toBe( true );
+        expect( fileContentRead ).toBe( fileContent );
     })
